move package metrics into /src/
