+++ src/package-metrics/metrics.ndjson
