+++ src/package-metrics/scripts/clean.sh
