+++ src/package-metrics/scripts/initial_setup.sh
