+++ src/package-metrics/scripts/setup.sh
