+++ src/package-metrics/src/index.ts
