+++ src/package-metrics/src/metric.ts
